bugfix:拷贝的excel被覆盖，添加新路径到path

+++ 考勤小程序/交付版本v2/run.py
@@ -2,6 +2,9 @@ import os.path
 import sys
 
 sys.path.append(os.path.abspath('.'))
+sys.path.append(os.path.abspath('./0-说明文档和配置文档'))
+sys.path.append(os.path.abspath('./1-学生名单表'))
+sys.path.append(os.path.abspath('./2-考勤结果'))
 sys.path.append(os.path.abspath('./考勤程序'))
 sys.path.append(os.path.abspath('./考勤程序/utils'))
 from 考勤程序 import server, service
@@ -12,13 +15,11 @@ TO_FILE = "考勤总表.xlsx"
 
 
 def init():
-    """
-    初始化：
-    复制文件
-    """
+    """ 初始化:复制文件 """
     source_file = service.get_read_file()
-    to_file = settings['save_folder'] + "/" + TO_FILE
-    file_utils.copy_file(source_file, to_file)
+    to_file = settings['save_folder'] + "/" + setting_utils.SAVE_FILE
+    if not os.path.exists(to_file):
+        file_utils.copy_file(source_file, to_file)
 
 
 if __name__ == '__main__':
